% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Single Trait Manhattan plot}
\usage{
manhattan_plot(mt_gwas_results, trait, bp_positions, ...)
}
\arguments{
\item{mt_gwas_results}{Results from function mt_gwas}

\item{trait}{integer number indicating the position of the trait to be plotted.}

\item{bp_positions}{dataframe with SNPs basepair positions. colnames should be 'chr' and 'position', rownames should be SNP identifiers matching names in mt_gwas.}

\item{...}{Further graphical parameters to customize the plot. Options include: title, bty, pch, cex.lab and/or cex.main.}
}
\description{
plots p-values with results from mt_gwas.
}
\author{
Original code by Fernando Aguate.
}
