% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_gwas.R
\name{mt_gwas}
\alias{mt_gwas}
\title{Multi-trait Genome wide association model.}
\usage{
mt_gwas(pleio_results, save_at = NULL)
}
\arguments{
\item{pleio_results}{object of class pleio_class (see function pleioR).}

\item{save_at}{string with directory and/or file name (.rdata) to save the results.}
}
\description{
Performs a multi-trait model with correlated errors (seemingly unrelated regressions), and generates results by trait.
}
\author{
Original code by Fernando Aguate.
}
