% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio_ideogram.R
\name{pleio_ideogram}
\alias{pleio_ideogram}
\title{Plot ideogram from pleio_test results}
\description{Plots the p-values or number of hits with the positions in the genome using results of pleio_test.}
\usage{
pleio_ideogram(
  pleio_res,
  alpha = "bonferroni05",
  n_traits = 2,
  bp_positions,
  window_size = 1e+06,
  centromeres = NULL,
  color_bias = 1,
  set_plot = T,
  set_legend = T,
  set_ylim_prop = 1.1,
  ...
)
}
