% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio_plot.R
\name{pleio_plot}
\alias{pleio_plot}
\title{Multi-trait manhattan plot}
\usage{
pleio_plot(
  pleio_res,
  alpha = "bonferroni05",
  n_traits = 2,
  bp_positions = NULL,
  set_colors = NULL,
  set_text = NULL,
  set_plot = TRUE,
  chr_spacing = 1e+05,
  ...
)
}
\arguments{
\item{pleio_res}{Object of class list that results of function pleio_test.}

\item{alpha}{either 'bonferroni05' or a numeric threshold for significance level.}

\item{n_traits}{number indicating the level of pleiotropy to plot.}

\item{bp_positions}{dataframe with chromosomes and basepair positions for SNPs matching pleio_res. rownames should contain SNP names.}

\item{set_colors}{string with 3 colors to use in the plot (by default: c('goldenrod4', 'brown4', 'royalblue2')).}

\item{set_text}{dataframe or matrix with text to add for some p-values. Row names should contain SNP names. Only the first column (with strings) will be considered.}

\item{set_plot}{boolean (TRUE by default) indicating whether to return the manhattan plot.}

\item{chr_spacing}{integer indicating the spacing (in basepair positions) between chromosomes. By default = 1e5.}

\item{...}{Additional graphic parameters for the plot.}
}
\description{
Plots the p-values that test the hypothesis of pleiotropic effects on n_traits.
}
\author{
Original code by Fernando Aguate.
}
