% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio_simulate.R
\name{pleio_simulate}
\alias{pleio_simulate}
\title{Create simulations}
\usage{
pleio_simulate(n_traits, n_individuals, n_snp, percentage_mv = 0)
}
\arguments{
\item{n_traits}{number of traits to simulate.}

\item{n_individuals}{number of individuals to simulate.}

\item{n_snp}{number of SNPs to simulate.}

\item{percentage_mv}{proportion of missing values. By default = 0.}
}
\value{
a list with pheno and geno to test the pleioR function.
}
\description{
Example function to create simulations with no effects.
}
\examples{
sim1 <- pleio_simulate(n_traits = 3, n_individuals = 1e4, n_snp = 1e3, percentage_mv = 0.1)
}
\author{
Original code by Fernando Aguate.
}
