% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiot.R
\docType{class}
\name{pleiot-class}
\alias{pleiot-class}
\title{Pleiotropy Testing Class}
\value{
An object of class `pleiot`.
}
\description{
Defines a class for storing pleiotropy testing results and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{pleio_object}}{An object of class `pleio`, which contains simulation details.}

\item{\code{pleiotest_stats}}{A `data.frame` containing pleiotropy test results for variants.}

\item{\code{execution_time}}{A `numeric` value indicating the time taken for execution.}
}}

