% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mtar.R
\name{run_cmats}
\alias{run_cmats}
\title{Run CMATS Pleiotropy Test}
\usage{
run_cmats(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` containing CMATS-derived p-values for each variant.
The row names correspond to variant identifiers.
}
\description{
Applies the `CMATS` method from the `MTARclone` package to test for pleiotropy using a covariance structure.
}
\details{
The function extracts the effect size (`beta`) and standard error (`SE`) matrices from the `pleio` object,
computes z-scores, and calculates their correlation structure.
The `CMATS` method is then used to estimate pleiotropy p-values.
}
\examples{
result <- run_cmats(pleio)
head(result)

}
