% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_dsum}
\alias{run_dsum}
\title{Run DSUM Pleiotropy Test}
\usage{
run_dsum(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `DSUM`-based p-values for each variant.
}
\description{
This function applies the `DSUM` method from the `MPAT` package to evaluate pleiotropy significance.
}
\details{
The function extracts effect size (`beta`) and standard error (`SE`) matrices from the `pleio` object,
computes z-scores, and estimates their correlation structure.
The `DSUM` method is then applied to compute p-values.
}
\examples{
result <- run_dsum(pleio)
head(result)

}
