% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mtar.R
\name{run_emats}
\alias{run_emats}
\title{Run EMATS Pleiotropy Test}
\usage{
run_emats(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with EMATS-derived p-values for each variant.
The row names correspond to variant identifiers.
}
\description{
Applies the `EMATS` method from the `MTARclone` package to estimate pleiotropy significance.
}
\details{
The function retrieves effect size (`beta`) and standard error (`SE`) matrices from the `pleio` object,
calculates z-scores, and estimates their correlation structure.
The `EMATS` method is then applied to test for pleiotropy.
}
\examples{
result <- run_emats(pleio)
head(result)

}
