% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_minp}
\alias{run_minp}
\title{Run MinP Pleiotropy Test}
\usage{
run_minp(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MinP`-based p-values for each variant.
}
\description{
Applies the `MinP` method from the `MPAT` package to assess pleiotropy.
}
\details{
This function extracts effect size (`beta`) and standard error (`SE`) matrices from the `pleio` object,
computes z-scores, and estimates their correlation structure.
The `MinP` method is then applied to compute the minimum p-value across phenotypes.
}
\examples{
result <- run_minp(pleio)
head(result)

}
