% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixada}
\alias{run_mixada}
\title{Run MixAda Pleiotropy Test}
\usage{
run_mixada(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `mixAda`-based p-values for each variant.
}
\description{
Applies the `mixAda` method from the `MPAT` package to assess pleiotropy.
}
\details{
Uses adaptive combination of association statistics across multiple traits.
}
\examples{
result <- run_mixada(pleio)
head(result)

}
