% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixfisher_liu}
\alias{run_mixfisher_liu}
\title{Run MixFisher Pleiotropy Test (Liu Approximation)}
\usage{
run_mixfisher_liu(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixFisher`-based p-values using the Liu method for each variant.
}
\description{
Applies the `MixFisher` method using the Liu approximation to assess pleiotropy.
}
\details{
Similar to `run_mixfisher_davies()`, but uses the Liu approximation instead of Davies.
}
\examples{
result <- run_mixfisher_liu(pleio)
head(result)

}
