% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixfisher_liumod}
\alias{run_mixfisher_liumod}
\title{Run MixFisher Pleiotropy Test (Liu-modified Approximation)}
\usage{
run_mixfisher_liumod(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixFisher`-based p-values using the Liu-modified method for each variant.
}
\description{
Applies the `MixFisher` method using the Liu-modified approximation to test for pleiotropy.
}
\details{
Uses a modified Liu approximation for estimating p-values in the `MixFisher` method.
}
\examples{
result <- run_mixfisher_liumod(pleio)
head(result)

}
