% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixsd_davies}
\alias{run_mixsd_davies}
\title{Run MixSD Pleiotropy Test (Davies Approximation)}
\usage{
run_mixsd_davies(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixSD`-based p-values using the Davies method for each variant.
}
\description{
Implements the `MixSD` test for pleiotropy using the Davies approximation.
}
\details{
The function applies the `MixSD` method to compute p-values.
}
\examples{
result <- run_mixsd_davies(pleio)
head(result)

}
