% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixsd_liu}
\alias{run_mixsd_liu}
\title{Run MixTippett Pleiotropy Test (Liu Approximation)}
\usage{
run_mixsd_liu(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixTippett`-based p-values using the Liu method for each variant.
}
\description{
Applies the `MixTippett` method using the Liu approximation to test for pleiotropy.
}
\details{
This function applies the `MixTippett` method using the Liu approximation.
}
\examples{
result <- run_mixtippett_liu(pleio)
head(result)

}
