% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixsd_liumod}
\alias{run_mixsd_liumod}
\title{Run MixSD Pleiotropy Test (Liu-modified Approximation)}
\usage{
run_mixsd_liumod(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixSD`-based p-values using the Liu-modified method for each variant.
}
\description{
Applies the `MixSD` method using the Liu-modified approximation to assess pleiotropy.
}
\details{
This function extracts effect size (`beta`) and standard error (`SE`) matrices from the `pleio` object,
computes z-scores, and estimates their correlation structure.
The `MixSD` method is then applied using the Liu-modified approximation to calculate p-values.
}
\examples{
result <- run_mixsd_liumod(pleio)
head(result)

}
