% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_mixvar_davies}
\alias{run_mixvar_davies}
\title{Run MixVar Pleiotropy Test (Davies Approximation)}
\usage{
run_mixvar_davies(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `MixVar`-based p-values using the Davies method for each variant.
}
\description{
Implements the `MixVar` test for pleiotropy using the Davies approximation.
}
\details{
The function applies the `MixVar` method to compute p-values using the Davies approximation.
}
\examples{
result <- run_mixvar_davies(pleio)
head(result)

}
