% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_pcfisher}
\alias{run_pcfisher}
\title{Run PC-Fisher Pleiotropy Test}
\usage{
run_pcfisher(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `PCFisher`-based p-values for each variant.
}
\description{
Applies the `PCFisher` method to evaluate pleiotropy significance.
}
\details{
This function extracts z-score matrices and applies the `PCFisher` method.
}
\examples{
result <- run_pcfisher(pleio)
head(result)

}
