% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mpat.R
\name{run_wald}
\alias{run_wald}
\title{Run Wald Pleiotropy Test}
\usage{
run_wald(pleio)
}
\arguments{
\item{pleio}{An object of class `pleio`, containing summary statistics for genetic variants.}
}
\value{
A `data.frame` with `Wald`-based p-values for each variant.
}
\description{
Applies the `Wald` test for pleiotropy.
}
\details{
The function extracts z-scores and applies the `Wald` method.
}
\examples{
result <- run_wald(pleio)
head(result)

}
