% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiot.R
\name{show,pleiot-method}
\alias{show,pleiot-method}
\title{Show Method for pleiot Class}
\usage{
\S4method{show}{pleiot}(object)
}
\arguments{
\item{object}{An object of class `pleiot`.}
}
\value{
Prints a summary of the pleiot object, including phenotypes, SNPs, and applied methods.
}
\description{
Prints a summary of the pleiot object.
}
\examples{
show(pleiotest)
}
