% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrsx_xrsy.R
\name{xrsx_xrsy}
\alias{xrsx_xrsy}
\title{Calculate XRsX and XRsY}
\usage{
xrsx_xrsy(id_matrix, sets_rs, xx, xy)
}
\arguments{
\item{id_matrix}{matrix of IDs}

\item{sets_rs}{list of inverses of matrix R}

\item{xx}{numeric vector with crossproducts of the X matrix}

\item{xy}{matrix with X transpose Y products.}
}
\description{
internal function to calculate crossproducts within pleioR.
}
