% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Single Trait Manhattan plot}
\usage{
manhattan_plot(mt_gwas_results, trait, bp_positions, ...)
}
\arguments{
\item{mt_gwas_results}{Object returned by mt_gwas}

\item{trait}{integer indicating the position of the trait (see: names(mt_gwas_results)) to be plotted.}

\item{bp_positions}{dataframe with SNPs base pair positions. colnames msut be 'chr' and 'position', rownames must be SNP identifiers matching names in mt_gwas.}

\item{...}{further graphical parameters. Options include: title=, bty=, pch=, cex.lab=, and cex.main=.}
}
\value{
Manhattan plot
}
\description{
Manhattan plot of results from mt_gwas function.
}
