% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_gwas.R
\name{mt_gwas}
\alias{mt_gwas}
\title{Multi-trait Genome wide association model.}
\usage{
mt_gwas(pleio_results, save_at = NULL)
}
\arguments{
\item{pleio_results}{object of class pleio_class (returned by pleioR() function).}

\item{save_at}{character with directory and/or file name (.rdata) to save the results. This is useful when handling multiple results such as in parallel jobs.}
}
\value{
list with by trait dataframes that contain results of the multi-trait model.
}
\description{
Performs a multi-trait model with correlated errors (seemingly unrelated regressions), and generates results by trait in a list.
}
