% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio_ideogram.R
\name{pleio_ideogram}
\alias{pleio_ideogram}
\title{Plot ideogram from pleio_test results}
\usage{
pleio_ideogram(
  pleio_res,
  alpha = "bonferroni05",
  n_traits = 2,
  bp_positions,
  window_size = 1e+06,
  centromeres = NULL,
  color_bias = 1,
  set_plot = T,
  set_legend = T,
  set_ylim_prop = 1.1,
  ...
)
}
\arguments{
\item{pleio_res}{Object of class list that results of function pleio_test.}

\item{alpha}{either 'bonferroni05' or a numeric threshold for significance level.}

\item{n_traits}{integer indicating the number of traits to be tested (not needed if type = "t").}

\item{bp_positions}{a data frame with column names 'chr' and 'pos' indicating the chromosome and position for each SNP. Row names must contain SNP names matching results from pleio_test.}

\item{window_size}{a numeric value indicating the size (in base pairs) of the genomic region that contains significant SNPs.}

\item{centromeres}{(default NULL) take string 'human' or a dataframe/matrix with chromosome in the first column and position (in mbp) for the centromere in the second column.}

\item{color_bias}{bias for the color scale. See colorRampPalette.}

\item{set_plot}{(default TRUE) plot the ideogram.}

\item{set_legend}{(default TRUE) plot legend.}

\item{set_ylim_prop}{(default = 1) proportion of upper margin to fit the legend, 1 = no margin, 1.1 = 10% left for margin}
}
\description{
Plots the p-values or number of hits with the positions in the genome using results of pleio_test.
}
\seealso{
\code{\link[pleio_plot]{pleio_plot}}
}
\author{
Original code by Fernando Aguate.
}
