% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio_test.R
\name{pleio_test}
\alias{pleio_test}
\title{Sequential Wald test for pleiotropy}
\usage{
pleio_test(pleio_results, loop_breaker = 0.99, save_at = NULL)
}
\arguments{
\item{pleio_results}{object of class pleio_class (see function pleioR).}

\item{loop_breaker}{A numerical value for a maximum p-value used to stop the sequence if they pass this threshold.}

\item{save_at}{string with directory and/or file name (.rdata) to save the results.}
}
\description{
Performs the sequential test to test pleiotropic effects from results of pleioR.
}
\author{
Original code by Fernando Aguate.
}
